+++ src/controllers/teacher.ts
@@ -1,5 +1,4 @@
 import { Context } from 'koa';
-import { identity } from 'lodash';
 import {TeacherService} from '../services/teacher'
 
 export default class TeacherControllers{
@@ -7,8 +6,16 @@ export default class TeacherControllers{
         const teacherService = new TeacherService()
         try {
             const {id,name,gender,subject} = ctx.request.body
-            const newTeacher = await teacherService.create({id,name,gender,subject})
-            ctx.body = newTeacher
+            ctx.checkBody('id').optional().notEmpty('teacher Id cannot be empty').isInt('subject id should be integer or number')
+            ctx.checkBody('name').len(3,20,'teacher name should be min of 3 characters and max of 20 characters ')
+            ctx.checkBody('gender').match(/m|M|Male|male|f|F|Female|female/,'gender should be male or female').notEmpty('gender cannot be empty')
+            ctx.checkBody('subject').len(3,20,'subject name should be min of 3 characters and max of 10 characters')
+            if(ctx.errors){
+                ctx.body = ctx.errors
+            }else{
+                const newTeacher = await teacherService.create({id,name,gender,subject})
+                ctx.body = newTeacher
+            }
         } catch (error) {
             console.log(error)
         }
@@ -39,8 +46,16 @@ export default class TeacherControllers{
         const teacherService = new TeacherService()
         try {
             const {id,name,gender,subject} = ctx.request.body;
-            const updatedTeacher = await teacherService.updateTeacher({id,name,gender,subject})
-            ctx.body = updatedTeacher
+            ctx.checkBody('id').optional().notEmpty('teacher Id cannot be empty').isInt('subject id should be integer or number')
+            ctx.checkBody('name').optional().len(3,20,'teacher name should be min of 3 characters and max of 20 characters ')
+            ctx.checkBody('gender').optional().match(/m|M|Male|male|f|F|Female|female/,'gender should be male or female').notEmpty('gender cannot be empty')
+            ctx.checkBody('subject').optional().len(3,20,'subject name should be min of 3 characters and max of 10 characters')
+            if(ctx.errors){
+                ctx.body = ctx.errors
+            }else{
+                const updatedTeacher = await teacherService.updateTeacher({id,name,gender,subject})
+                ctx.body = updatedTeacher
+            }
         } catch (error) {
             console.log(error)
         }
@@ -50,8 +65,14 @@ export default class TeacherControllers{
         const teacherService = new TeacherService()
         try {
             const {id} = ctx.request.body;
-            const deletedTeacher = await teacherService.deleteTeacher(id);
-            ctx.body = "deleted Succesfully"
+            ctx.checkBody('id').notEmpty('id cannot be null').isInt('id should be number ')
+            if(ctx.errors){
+                ctx.body = ctx.errors
+            }
+            else{
+                const deletedTeacher = await teacherService.deleteTeacher(id);
+                ctx.body = "deleted Succesfully"
+            }
         } catch (error) {
             console.log(error)
         }
